% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_server.R
\name{analysis_server}
\alias{analysis_server}
\title{Analyse data}
\usage{
analysis_server(id, course_paths, references)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{references}{Reactive. Function containing a list of references.}
}
\value{
A quarto document which can be rendered in different formats.
}
\description{
Module allowing the user to analyse data about teaching, learning, and testing materials.
}
\author{
Nicolas Mangin
}
