% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_ui.R
\name{analysis_ui}
\alias{analysis_ui}
\title{Analyse data}
\usage{
analysis_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A quarto document which can be rendered in different formats.
}
\description{
Module allowing the user to analyse data about teaching, learning, and testing materials.
}
\author{
Nicolas Mangin
}
