% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_selections_to_analysis.R
\name{apply_selections_to_analysis}
\alias{apply_selections_to_analysis}
\title{Change selections and filters}
\usage{
apply_selections_to_analysis(selections, analysis_path)
}
\arguments{
\item{selections}{Character vector. Selected tree, language, student, test, and question.}

\item{analysis_path}{Character. Path to the folder containing the analysis.}
}
\description{
Function updating the relevant selections in the selected quarto document.
}
\author{
Nicolas Mangin
}
