% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback_ui.R
\name{feedback_ui}
\alias{feedback_ui}
\title{Edit feedback}
\usage{
feedback_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\description{
Module facilitating the quick creation or modification of individual feedback.
}
\author{
Nicolas Mangin
}
