% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_report}
\alias{html_report}
\title{HTML report template for Rmarkdown}
\usage{
html_report(...)
}
\arguments{
\item{...}{further arguments pssed to rmarkdown::html_document()}
}
\description{
HTML report template for Rmarkdown
}
