% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_out.R
\name{lin_out}
\alias{lin_out}
\title{Find Outliers in a Linear Model}
\usage{
lin_out(y, x, deg)
}
\arguments{
\item{y}{A numerical vector as the independent variable}

\item{x}{A numerical vector as the dependent variable}

\item{deg}{An integer determining the degree of the linear model}
}
\value{
A table of the vertical distances of each y-value from its predicted y-value
}
\description{
Find Outliers in a Linear Model
}
\examples{
x1 <- c(4, 8, 2, 0, 1, 5, 3, 9, 2, 7, 3, 5, 4, 10, 4)
y1 <- c(5, 2, 7, 5, 1, 5, 2, 6, 9, 1, 6, 3, 7, 8, 41)
output <- lin_out(y1, x1, 1)
}
