% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_out.R
\name{log_out}
\alias{log_out}
\title{Find Outliers in a Logistic Model}
\usage{
log_out(y, x, deg, positive)
}
\arguments{
\item{y}{A binomial categorical vector as the independent variable}

\item{x}{A numerical vector as the dependent variable}

\item{deg}{An integer determining the degree of the logistic model}

\item{positive}{A character value that determines which category in y will be considered "1" rather than "0"}
}
\value{
A table of the vertical distances of each y-value (0 or 1) from its predicted probability
}
\description{
Find Outliers in a Logistic Model
}
\examples{
x1 <- c(70, 81, 22, 65, 21, 32, 7, 71, 58, 37, 52, 83, 4, 71, 97)
y1 <- c("TRUE", "TRUE", "FALSE", "TRUE", "FALSE", "FALSE", "FALSE", "FALSE",
        "TRUE", "FALSE", "FALSE", "TRUE", "FALSE", "TRUE", "FALSE")
output <- log_out(y1, x1, 1, "TRUE")
}
