% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{{toy_data}}
\alias{{toy_data}}
\alias{toy_data}
\title{toy_dataset for reportr purposes}
\format{
A data frame with 15 rows and 3 variables:
\describe{
\item{x1}{A numerical column that can operate as the sole predictor required for lin_out and log_out}
\item{y1}{A numerical column that can operate as a numerical response required for lin_out}
\item{y2}{A categorical column that can operate as a binary response required for log_out}
}
}
\usage{
toy_data
}
\description{
A 3x15 toy dataset that allows the user to familiarize themselves with lin_out, log_out, and kick_out.
}
\keyword{datasets}
