% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{PrettyTable}
\alias{PrettyTable}
\title{Create prettier HTML tables}
\usage{
PrettyTable(DT, height = "400px", width = "900px", addScroll = FALSE)
}
\arguments{
\item{DT}{A data.frame or data.table.}

\item{height}{character. Passed to `kableExtra::scroll_box`. Will only be
used if `addScroll = TRUE`.}

\item{width}{character. Passed to `kableExtra::scroll_box`. Will only be
used if `addScroll = TRUE`.}

\item{addScroll}{boolean. Controls whether to add a scrollbar to the table.
Note that `height` and `width` fix the table to that exact width, there is
no autosizing for short tables, so only only use `addScroll` on large tables.}
}
\description{
A quick wrapper to simplify the presentation of kable tables
for this project.
}
