% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_for_analysis.R
\name{retrieve_data_for_analysis}
\alias{retrieve_data_for_analysis}
\title{Analyse data}
\usage{
retrieve_data_for_analysis(analysis_path, course_paths)
}
\arguments{
\item{analysis_path}{Character. Path to the folder containing the analysis.}

\item{course_paths}{List. List of paths to the different folders and databases on local disk.}
}
\value{
A quarto document which can be rendered in different formats.
}
\description{
Function copying and pasting required databases in the data folder of the selected analysis.
}
\author{
Nicolas Mangin
}
