% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{VarSum}
\alias{VarSum}
\title{Create a metadata summary of a table.}
\usage{
VarSum(df, exclude = "values")
}
\arguments{
\item{df}{A data.frame or data.table.}

\item{exclude}{character. A vector to exclude certain columns from view.}
}
\description{
Create a metadata summary of a table.
}
