% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print_meta}
\alias{print_meta}
\title{HTML report template for Rmarkdown}
\usage{
print_meta(Statistician = "me <myfault@nohelp.com>",
  Collaborator = "you <yourfault@nohelp.com>", ...)
}
\arguments{
\item{...}{further arguments pssed to rmarkdown::html_document()}
}
\description{
HTML report template for Rmarkdown
}
