% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_detailed_feeback.R
\name{provide_detailed_feeback}
\alias{provide_detailed_feeback}
\title{Table providing feedback on closed questions}
\usage{
provide_detailed_feeback(
  details,
  studentid,
  labels = base::list(values = c("Wrong", "Missing", "Partial", "Correct"), table_names =
    c("Proposition:", "This proposition is:", "Therefore your answer is:",
    "And you earned:", "Explanation", "Topic"))
)
}
\arguments{
\item{details}{Tibble. Feedback recorded for all items of all versions of all questions.}

\item{studentid}{Character. ID of the student.}

\item{labels}{List. Names for axes and legends (for translations purposes)}
}
\value{
HTML code Individual feedback properly formatted
}
\description{
Function formatting the feedback about multiple choice questions into a HTML table.
}
\author{
Nicolas Mangin
}
