% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{PDF report template for RMarkdown}
\usage{
report(...)
}
\arguments{
\item{...}{further arguments pssed to rmarkdown::pdf_document()}
}
\description{
PDF report template for RMarkdown
}
