\name{reportR-package}
\alias{reportR-package}
\alias{reportR}
\docType{package}
\title{
\packageTitle{reportR}
}
\description{
\packageDescription{reportR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{reportR}
\packageIndices{reportR}
This package consists of of small miscellaneous functions to streamline the process of compiling attractive and consistently formatted knitr reports, including a ggplot2 theme (\code{\link{custom_theme}}), a CSS file to sylize knitr documents (\code{\link{sink_css}}), and a function to stylize the headers of knitr documents \code{\link{print_header}}.
}
\author{
\packageAuthor{reportR}

Maintainer: \packageMaintainer{reportR}
}
\keyword{ knitr, report, css, ggplot2 }
  \seealso{
    \code{\link{custom_theme}}
    \code{\link{print_header}}
    \code{\link{sink_css}}
  }
\examples{

}
