\name{reportR-package}
\alias{reportR-package}
\alias{reportR}
\docType{package}
\title{
\packageTitle{reportR}
}
\description{
\packageDescription{reportR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{reportR}
\packageIndices{reportR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{reportR}

Maintainer: \packageMaintainer{reportR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
